feat: 🎸 TODO-73: Refactor some codes
