+++ src/pages/action/index.tsx
@@ -8,8 +8,7 @@ import Button from '@/core-ui/button';
 import Input from '@/core-ui/input';
 
 import styles from './style.module.scss';
-import {useEffect} from 'react';
-import {ROUTES} from '@/configs/routes.config';
+import useCheckUserLocalStorage from '@/hooks/useCheckUserLocalStorage';
 
 interface IFormInputs {
   ID: string;
@@ -20,13 +19,8 @@ const Schema = yup.object().shape({
 });
 const Action: React.FC = () => {
   const router = useRouter();
-  // Check local storage.
-  // useEffect(() => {
-  //   const checkLocal = localStorage.getItem('user');
-  //   if (!checkLocal) {
-  //     router.push(ROUTES.QUICKPLAY);
-  //   }
-  // }, []);
+
+  useCheckUserLocalStorage();
 
   const {register, handleSubmit, formState} = useForm<IFormInputs>({
     mode: 'onChange',
@@ -34,8 +28,6 @@ const Action: React.FC = () => {
   });
 
   const {errors} = formState;
-
-  useEffect(() => {});
 
   const onSubmit: SubmitHandler<IFormInputs> = data => {
     API.readTodoList(Number(data.ID))
@@ -70,7 +62,7 @@ const Action: React.FC = () => {
                 </Button>
                 <form className="input-group" onSubmit={handleSubmit(onSubmit)}>
                   <Input className={errors.ID && 'error'} placeholder="Enter ID" {...register('ID')} />
-                  <Button className="input-group-text" type="submit">
+                  <Button variant="outlined" className="input-group-text" type="submit">
                     Join
                   </Button>
                 </form>
